refactor: move spinner to dedicated module
